Emulate Flask's import_path/root_path model to simplify resource loading relative to graph.
In a nutshell: we want to use the name of the application as a path to find resources.

We can't just assume the name is a module name, so allow the user to pick an import
name or even set an absolute root path. But don't go to quite the extremes Flask does;
if the user doesn't want to load resources relative to the graph, don't make them.