Isolate hooks between class and subclass
**Why?**
Component hooks are registered on the class that will be instantiated for the component, through a class attribute.
This means that if component `a` is an instance of `class A`, and component `b` an instance of `class B(A)` then the hooks defined on those two components will get mixed.

**What?**
Ensure separation of hooks in cases like the one above by using the class name in the class attribute used to hold hooks.