Fix boolean typed config
**Why?**
In many cases configuration is passed to the application as a string (e.g. environment variable). We use the `typed` function to convert those strings to the proper type.
Today that type conversion is a simple cast, meaning that `"False"` gets cast to `True`.
This hadn't surfaced before because we only had cases where the default value was `False`, and it would configured to `"True"` via external config.

**What?**
- Use existing `boolean` function for proper type conversion.
- Fix `check_no_warnings` decorator.